refactor (#427)
- prepare for navigation
