+++ git-revision/src/spec/parse.rs
@@ -31,34 +31,22 @@ pub enum Error {
 pub mod delegate {
     use bstr::BStr;
 
-    /// Usually the first methods to call when parsing a rev-spec to set an anchor.
+    /// Usually the first methods to call when parsing a rev-spec to set an anchoring revision (which is typically a `Commit` object).
     /// Methods can be called multiple time to either try input or to parse another rev-spec that is part of a range.
     ///
     /// In one case they will not be called at all, e.g. `@{[-]n}` indicates the current branch (what `HEAD` dereferences to),
     /// without ever naming it, and so does `@{upstream}` or `@{<date>}`.
-    pub trait Anchor {
+    ///
+    /// Note that when dereferencing `HEAD` implicitly, a revision must be set for later navigation.
+    pub trait Revision {
         /// Resolve `name` as reference which might not be a valid reference name. The name may be partial like `main` or full like
         /// `refs/heads/main` solely depending on the users input.
         /// Symbolic referenced should be followed till their object, but objects must not yet be peeled.
         fn find_ref(&mut self, name: &BStr) -> Option<()>;
+
         /// An object prefix to disambiguate, returning `None` if it is ambiguous or wasn't found at all.
         fn disambiguate_prefix(&mut self, prefix: git_hash::Prefix) -> Option<()>;
-    }
 
-    /// Combine one or more specs into a range of multiple.
-    pub trait Kind {
-        /// Set the kind of the spec, which happens only once if it happens at all.
-        /// In case this method isn't called, assume `Single`.
-        /// Reject a kind by returning `None` to stop the parsing.
-        ///
-        /// Note that ranges don't necessarily assure that a second specification will be parsed.
-        /// If `^rev` is given, this method is called with [`spec::Kind::Range`][crate::spec::Kind::Range]
-        /// and no second specification is provided.
-        fn kind(&mut self, kind: crate::spec::Kind) -> Option<()>;
-    }
-
-    /// Once an anchor is set one can adjust it using navigation methods.
-    pub trait Navigation {
         /// Lookup the reflog of the previously set reference, or dereference `HEAD` to its reference
         /// to obtain the ref name (as opposed to `HEAD` itself).
         /// If there is no such reflog entry, return `None`.
@@ -73,20 +61,47 @@ pub mod delegate {
         /// in the configuration (typically in `refs/remotes/…`). The sibling branches are always local tracking branches.
         /// Return `None` of no such configuration exists and no sibling could be found, which is also the case for all reference outside
         /// of `refs/heads/`.
-        /// Note that the caller isn't aware if the previously set reference is a branch or not.
+        /// Note that the caller isn't aware if the previously set reference is a branch or not and might call this method even though no reference
+        /// is known.
         fn sibling_branch(&mut self, kind: SiblingBranch) -> Option<()>;
     }
 
+    /// Combine one or more specs into a range of multiple.
+    pub trait Kind {
+        /// Set the kind of the spec, which happens only once if it happens at all.
+        /// In case this method isn't called, assume `Single`.
+        /// Reject a kind by returning `None` to stop the parsing.
+        ///
+        /// Note that ranges don't necessarily assure that a second specification will be parsed.
+        /// If `^rev` is given, this method is called with [`spec::Kind::Range`][crate::spec::Kind::Range]
+        /// and no second specification is provided.
+        fn kind(&mut self, kind: crate::spec::Kind) -> Option<()>;
+    }
+
+    /// Once an anchor is set one can adjust it using traversal methods.
+    pub trait Navigate {
+        /// Adjust the current revision to traverse the graph according to `kind`.
+        fn traverse(&mut self, kind: Traversal) -> Option<()>;
+    }
+
     /// A lookup into the reflog of a reference.
-    #[derive(Debug, Copy, Clone)]
+    #[derive(PartialEq, Eq, Debug, Hash, Ord, PartialOrd, Clone, Copy)]
     pub enum ReflogLookup {
         /// Lookup by entry, where `0` is the most recent entry, and `1` is the older one behind `0`.
         Entry(usize),
         Date(git_date::Time),
     }
 
+    /// Define how to traverse the commit graph.
+    #[derive(PartialEq, Eq, Debug, Hash, Ord, PartialOrd, Clone, Copy)]
+    pub enum Traversal {
+        /// Select the given parent commit of the currently selected commit, start at `1` for the first parent.
+        /// The value will never be `0`.
+        NthParent(usize),
+    }
+
     /// The kind of sibling branch to obtain.
-    #[derive(Debug, Copy, Clone)]
+    #[derive(PartialEq, Eq, Debug, Hash, Ord, PartialOrd, Clone, Copy)]
     pub enum SiblingBranch {
         /// The upstream branch as configured in `branch.<name>.remote` or `branch.<name>.merge`.
         Upstream,
@@ -112,9 +127,9 @@ pub mod delegate {
 /// - **Anchors** - which revision to use as starting point for…
 /// - **Navigation** - where to go once from the initial revision
 /// - **Range** - to learn if the specification is for a single or multiple references, and how to combine them.
-pub trait Delegate: delegate::Anchor + delegate::Navigation + delegate::Kind {}
+pub trait Delegate: delegate::Revision + delegate::Navigate + delegate::Kind {}
 
-impl<T> Delegate for T where T: delegate::Anchor + delegate::Navigation + delegate::Kind {}
+impl<T> Delegate for T where T: delegate::Revision + delegate::Navigate + delegate::Kind {}
 
 pub(crate) mod function {
     use crate::spec;
